Implemented gradual decline points given to the user when multiple movies are added in a single day.
Implemented the try-hard mode, where the user must wait the full movie's runtime to receive points or interact with the added movie (rate it or delete it). This means that he also needs the full runtime to end so that he can add a new movie.

Fixed some small bugs while implementing those big features.